Move binary prefixes.
