--- lib/alchemist/binary_prefixes.rb
@@ -1,25 +1,0 @@
-module Alchemist
-  def self.binary_prefixes
-    {
-      :yotta => 2.0**80.0, :Y => 2.0**80,
-      :zetta => 2.0**70.0, :Z => 2.0**70.0,
-      :exa => 2.0**60.0, :E => 2.0**60.0,
-      :peta => 2.0**50.0, :P => 2.0**50.0,
-      :tera => 2.0**40.0, :T => 2.0**40.0,
-      :giga => 2.0**30.0, :G => 2.0**30.0,
-      :mega => 2.0**20.0, :M => 2.0**20.0,
-      :kilo => 2.0**10.0, :k => 2.0**10.0,
-
-      # binary prefixes
-
-      :kibi => 2.0**10.0, :Ki => 2.0**10.0,
-      :mebi => 2.0**20.0, :Mi => 2.0**20.0,
-      :gibi => 2.0**30.0, :Gi => 2.0**30.0,
-      :tebi => 2.0**40.0, :Ti => 2.0**40.0,
-      :pebi => 2.0**50.0, :Pi => 2.0**50.0,
-      :exbi => 2.0**60.0, :Ei => 2.0**60.0,
-      :zebi => 2.0**70.0, :Zi => 2.0**70.0,
-      :yobi => 2.0**80.0, :Yi => 2.0**80.0
-    }
-  end
-end
